+++ classes/Ergo/Http/Client.php
@@ -10,18 +10,17 @@ use Ergo\Http\Error;
 class Client
 {
 	const MAX_REDIRECTS=10;
-	const DEFAULT_TIMEOUT=10;
 
 	private $_url;
 	private $_redirects=0;
 	private $_filters=array();
 	private $_headers=array();
-	private $_proxy;
-	private $_auth;
-	private $_timeout=self::DEFAULT_TIMEOUT;
+
 
 	public static $requestCount=0;
 	public static $requestTime=0;
+
+	private static $_transport;
 
 	/**
 	 * @param string $url
@@ -30,6 +29,17 @@ class Client
 	{
 		if (!$url) throw new \InvalidArgumentException('A base url must be set');
 		$this->_url = new Url($url);
+	}
+
+	public static function transport($transport=null)
+	{
+		if (!is_null($transport))
+			self::$_transport = $transport;
+
+		if (!isset(self::$_transport))
+			self::$_transport = new Transport();
+
+		return self::$_transport;
 	}
 
 	/**
@@ -59,7 +69,7 @@ class Client
 	 */
 	public function setHttpProxy($url)
 	{
-		$this->_proxy = $url;
+		self::transport()->setHttpProxy($url);
 		return $this;
 	}
 
@@ -68,7 +78,17 @@ class Client
 	 */
 	public function setHttpAuth($user, $pass)
 	{
-		$this->_auth = $user . ':' . $pass;
+		self::transport()->setHttpAuth();
+		return $this;
+	}
+
+	/**
+	 * Sets the connection timeout in seconds
+	 * @chainable
+	 */
+	public function setTimeout($seconds)
+	{
+		self::transport()->setTimeout($seconds);
 		return $this;
 	}
 
@@ -100,7 +120,9 @@ class Client
 	 */
 	function get($path)
 	{
-		return $this->_dispatchRequest($this->_buildRequest('GET',$path));
+		return $this->_dispatchRequest(
+			$this->_buildRequest('GET', $path)
+		);
 	}
 
 	/**
@@ -109,37 +131,9 @@ class Client
 	 */
 	function delete($path)
 	{
-		return $this->_dispatchRequest($this->_buildRequest('DELETE',$path));
-	}
-
-	/**
-	 * Parses a response into headers and a body
-	 */
-	private function _buildResponse($response)
-	{
-		$sections = explode("\r\n\r\n", $response,2);
-		$body = isset($sections[1]) ? $sections[1] : NULL;
-		$headers = array();
-		$headerlines = explode("\n",$sections[0]);
-
-		// process status
-		list($http, $code, $message) = explode(' ',$headerlines[0],3);
-
-		// process headers
-		foreach(array_slice($headerlines,1) as $headerline)
-		{
-			$headers[] = HeaderField::fromString($headerline);
-		}
-
-		$response = new Response($code,$headers,$body);
-
-		// pass the response through the filter chain
-		foreach($this->_filters as $filter)
-		{
-			$response = $filter->response($response);
-		}
-
-		return $response;
+		return $this->_dispatchRequest(
+			$this->_buildRequest('DELETE',$path)
+		);
 	}
 
 	/**
@@ -161,6 +155,12 @@ class Client
 			$body
 		);
 
+		// pass the request through the filter chain
+		foreach($this->_filters as $filter)
+		{
+			$request = $filter->request($request);
+		}
+
 		return $request;
 	}
 
@@ -169,30 +169,21 @@ class Client
 	 */
 	private function _dispatchRequest($request)
 	{
-		// pass the request through the filter chain
-		foreach($this->_filters as $filter)
-		{
-			$request = $filter->request($request);
-		}
-
 		// track the number of requests across instances
 		self::$requestCount++;
 		$timestart = microtime(true);
 
-		// prepare and send the curl request
-		$curl = $this->_curlConnection($request);
-		if(($curlResponse = curl_exec($curl)) === false)
+		$response = self::transport()->send($request);
+
+		// pass the response through the filter chain
+		foreach($this->_filters as $filter)
 		{
-			throw new Error('Curl error: ' . curl_error($curl),
-				curl_errno($curl));
+			$response = $filter->response($response);
 		}
 
-		$response = $this->_buildResponse($curlResponse);
 		$httpCode = $response->getStatus()->getCode();
 		$location = $response->getHeaders()->value('Location');
 		$body = $response->getBody();
-
-		curl_close($curl);
 
 		// track the time taken across instances
 		self::$requestTime += microtime(true) - $timestart;
@@ -233,62 +224,6 @@ class Client
 	}
 
 	/**
-	 * Initializes the curl connection
-	 */
-	private function _curlConnection($request)
-	{
-		// create a new curl resource
-		$curl = curl_init();
-		$method = $request->getRequestMethod();
-		$headers = array('Expect:');
-
-		// add existing headers into a flat string format
-		foreach($request->getHeaders() as $header)
-		{
-			$headers[] = rtrim($header->__toString());
-		}
-
-		// set URL and other appropriate options
-		curl_setopt($curl, CURLOPT_URL, $request->getUrl());
-		curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
-		curl_setopt($curl, CURLOPT_HEADER, true);
-		curl_setopt($curl, CURLOPT_VERBOSE, false);
-		curl_setopt($curl, CURLOPT_TIMEOUT, $this->_timeout);
-		curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, false);
-		curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, false);
-
-		// enable proxy support
-		if(isset($this->_proxy))
-		{
-			curl_setopt($curl, CURLOPT_PROXY, $this->_proxy);
-		}
-
-		// enable http authentication
-		if(isset($this->_auth))
-		{
-			curl_setopt($curl, CURLOPT_USERPWD, $this->_auth);
-			curl_setopt($curl, CURLOPT_HTTPAUTH, CURLAUTH_BASIC);
-		}
-
-		if($method == 'PUT' || $method == 'POST')
-		{
-			$headers[] = 'Content-Length: '.strlen($request->getBody());
-
-			curl_setopt($curl, CURLOPT_CUSTOMREQUEST, $method);
-			curl_setopt($curl, CURLOPT_POSTFIELDS, $request->getBody());
-		}
-		elseif($method == 'DELETE')
-		{
-			curl_setopt($curl, CURLOPT_CUSTOMREQUEST, $method);
-		}
-
-		// add HTTP headers
-		curl_setopt($curl, CURLOPT_HTTPHEADER, $headers);
-
-		return $curl;
-	}
-
-	/**
 	 * Redirect to a new url
 	 */
 	private function _redirect($location)
@@ -310,15 +245,5 @@ class Client
 
 		return $this->_dispatchRequest(
 			new Request('GET', $locationUrl, $this->_headers));
-	}
-
-	/**
-	 * Sets the connection timeout in seconds
-	 * @chainable
-	 */
-	public function setTimeout($seconds)
-	{
-		$this->_timeout = $seconds;
-		return $this;
 	}
 }
